Alter feeds query for greater efficacy+efficiency.
You can't use $post that way, unless you declare it as an object WordPress doesn't give it scope inside The Loop. That is an empty object. 

Also feedUrl isn't an aspect of the object, we couldn't set it and it can't be accessed that way even if we did declare that as a piece of post_meta.

We store the URL of each feed in the GUID aspect though, so it is directly accessible from the post object and because it is a unique identifier. 

With this in mind, you can use the standard functions to access that property in the loop. And as long as PressForward is on, the post_type variable should be accessible via this method.