Perhaps we shouldn't be relying on GUID
It occurs to me, as we look over this and other places we deal with feeds in the WordPress system, that accessing the feed URL via GUID may not be the way to go, because perhaps we shouldn't be storing it there. What if we actually want to treat our feeds like posts and make a list of feeds with descriptions of them that people can browse and page through? I think having the GUID set with the URL would make this a mess. As our intent changes with the feeds, this may no longer make sense. So let's access it via post meta instead.